ZD3584380: handle expired single idp cookie on continue-to-your-idp page
Move cookie checking code to single_idp_partial_controller as it is now used in two controllers.
Turn off caching for pages generated by single idp controller so that the proper page is shown if the user clicks back after cookie has expired.
