+++ app/controllers/redirect_to_idp_controller.rb
@@ -1,7 +1,9 @@
 require 'partials/idp_selection_partial_controller'
+require 'partials/single_idp_partial_controller'
 
 class RedirectToIdpController < ApplicationController
   include IdpSelectionPartialController
+  include SingleIdpPartialController
 
   def register
     request_form
@@ -32,12 +34,16 @@ class RedirectToIdpController < ApplicationController
   end
 
   def single_idp
-    uuid = MultiJson.load(cookies.encrypted[CookieNames::VERIFY_SINGLE_IDP_JOURNEY]).fetch('uuid', nil)
-    request_form_for_single_idp_journey(uuid)
-    increase_attempt_number
-    report_user_idp_attempt_to_piwik
-    FEDERATION_REPORTER.report_single_idp_journey_selection(current_transaction, request, session[:selected_idp_name])
-    render :redirect_to_idp
+    if valid_cookie?
+      uuid = MultiJson.load(cookies.encrypted[CookieNames::VERIFY_SINGLE_IDP_JOURNEY]).fetch('uuid', nil)
+      request_form_for_single_idp_journey(uuid)
+      increase_attempt_number
+      report_user_idp_attempt_to_piwik
+      FEDERATION_REPORTER.report_single_idp_journey_selection(current_transaction, request, session[:selected_idp_name])
+      render :redirect_to_idp
+    else
+      render_error 'session_timeout', :forbidden
+    end
   end
 
 private
